hh/requirements.txt
flask
google-generativeai
PyMuPDF
python-dotenv
